Initial Commit
This is the initial commit of the swerve test bot. The goal is to first
get basic functionality going, and then try some fun swerve optimization
stuff. SwerveSubsystem needs to be fleshed out, and the structure of the
project should be maintained in such a way to where we can do the
optimization things we want to.